# docs/PRODUCT/openapi.yaml
openapi: 3.1.0
info:
  title: Pazar Marketplace API (Canonical)
  version: 1.4.0
  description: >
    Canonical OpenAPI contract for the Pazar service (world: marketplace).
    This spec intentionally does NOT describe any retired multiworld routing such as /api/v1/{world}/...
servers:
  - url: http://localhost:8080
...
paths:
  /api/world/status:
    get:
      tags: [World]
      summary: World status for Pazar (marketplace)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorldStatus"
        default:
          $ref: "#/components/responses/ErrorEnvelope"
...
components:
...
  responses:
    ErrorEnvelope:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      required: [request_id, error]
      properties:
        request_id:
          type: string
          description: Correlation id for tracing/debugging.
        error:
          $ref: "#/components/schemas/ErrorObject"
        details:
          type: object
          additionalProperties: true
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    WorldStatus:
      type: object
      required: [world_key, availability, phase, version]
      properties:
        world_key:
          type: string
          example: marketplace
        availability:
          type: string
          example: ONLINE
        phase:
          type: string
          example: GENESIS
        version:
          type: string
...